refactor(schemas): move auth tokens to schemas/auth-tokens/
Per-dialect Drizzle table definitions for verificationTokens,
emailVerificationTokens, passwordResetTokens, and refreshTokens now live in
schemas/auth-tokens/{postgres,mysql,sqlite}.ts with a dialect-aware barrel
exposing authTokenTables(dialect).

database/schema/<dialect>.ts re-exports from the new location so existing
imports and the refreshTokensRelations() block keep compiling. Relations
themselves stay in database/schema/<dialect>.ts during Plan A and
consolidate in Task 17.
