+++ packages/nextly/src/database/schema/sqlite.ts
@@ -22,21 +22,21 @@ export const systemMigrations = sqliteTable("system_migrations", {
 // until Task 16 sweeps imports to @nextly/schemas.
 export { users, accounts, sessions } from "../../schemas/users/sqlite";
 
-export const verificationTokens = sqliteTable(
-  "verification_tokens",
-  {
-    identifier: text("identifier").notNull(),
-    token: text("token").notNull(),
-    expires: integer("expires", { mode: "timestamp" }).notNull(),
-  },
-  t => [
-    uniqueIndex("verification_tokens_identifier_token_pk").on(
-      t.identifier,
-      t.token
-    ),
-    index("verification_tokens_token_idx").on(t.token),
-  ]
-);
+// Auth-token tables — moved to schemas/auth-tokens/sqlite.ts (Plan A Task 6).
+// Re-exported here so existing consumers and relations() blocks below keep working
+// until Task 16 sweeps imports to @nextly/schemas.
+export {
+  verificationTokens,
+  emailVerificationTokens,
+  passwordResetTokens,
+  refreshTokens,
+} from "../../schemas/auth-tokens/sqlite";
+import {
+  verificationTokens,
+  emailVerificationTokens,
+  passwordResetTokens,
+  refreshTokens,
+} from "../../schemas/auth-tokens/sqlite";
 
 // Audit table for dynamic DDL
 export const contentSchemaEvents = sqliteTable(
@@ -54,76 +54,6 @@ export const contentSchemaEvents = sqliteTable(
   t => [
     index("content_schema_events_created_at_idx").on(t.createdAt),
     index("content_schema_events_table_name_idx").on(t.tableName),
-  ]
-);
-
-// Password reset tokens (custom table)
-export const passwordResetTokens = sqliteTable(
-  "password_reset_tokens",
-  {
-    id: integer("id").primaryKey({ autoIncrement: true }),
-    identifier: text("identifier").notNull(),
-    tokenHash: text("token_hash").notNull(),
-    expires: integer("expires", { mode: "timestamp" }).notNull(),
-    usedAt: integer("used_at", { mode: "timestamp" }),
-    createdAt: integer("created_at", { mode: "timestamp" })
-      .notNull()
-      .$defaultFn(() => new Date()),
-  },
-  t => [
-    uniqueIndex("prt_identifier_token_hash_unique").on(
-      t.identifier,
-      t.tokenHash
-    ),
-    index("prt_expires_idx").on(t.expires),
-    index("prt_used_at_idx").on(t.usedAt),
-  ]
-);
-
-// Email verification tokens (custom, hashed)
-export const emailVerificationTokens = sqliteTable(
-  "email_verification_tokens",
-  {
-    id: integer("id").primaryKey({ autoIncrement: true }),
-    identifier: text("identifier").notNull(),
-    tokenHash: text("token_hash").notNull(),
-    expires: integer("expires", { mode: "timestamp" }).notNull(),
-    createdAt: integer("created_at", { mode: "timestamp" })
-      .notNull()
-      .$defaultFn(() => new Date()),
-  },
-  t => [
-    uniqueIndex("evt_identifier_token_hash_unique").on(
-      t.identifier,
-      t.tokenHash
-    ),
-    index("evt_expires_idx").on(t.expires),
-  ]
-);
-
-// Refresh tokens for custom auth session management
-// Stores SHA-256 hashed opaque tokens, enables session revocation and token rotation
-export const refreshTokens = sqliteTable(
-  "refresh_tokens",
-  {
-    id: text("id").primaryKey(),
-    userId: text("user_id")
-      .notNull()
-      .references(() => users.id, { onDelete: "cascade" }),
-    // SHA-256 hex digest of the opaque refresh token (never store raw tokens)
-    tokenHash: text("token_hash").notNull(),
-    // Request metadata for session listing and security auditing
-    userAgent: text("user_agent"),
-    ipAddress: text("ip_address"),
-    expiresAt: integer("expires_at", { mode: "timestamp" }).notNull(),
-    createdAt: integer("created_at", { mode: "timestamp" })
-      .notNull()
-      .$defaultFn(() => new Date()),
-  },
-  t => [
-    index("refresh_tokens_token_hash_idx").on(t.tokenHash),
-    index("refresh_tokens_user_id_idx").on(t.userId),
-    index("refresh_tokens_expires_at_idx").on(t.expiresAt),
   ]
 );
