+++ packages/nextly/src/schemas/auth-tokens/sqlite.ts
@@ -1,0 +1,112 @@
+/**
+ * Auth-token tables — SQLite.
+ *
+ * Tables: verificationTokens, emailVerificationTokens, passwordResetTokens,
+ * refreshTokens.
+ * Moved verbatim from packages/nextly/src/database/schema/sqlite.ts as part of
+ * Plan A schemas consolidation. No behavior change.
+ *
+ * Note: cross-table `relations()` blocks (refreshTokensRelations) remain in
+ * database/schema/sqlite.ts during Plan A — they reference tables that move
+ * in later tasks. Relations consolidate in Task 17 once database/schema/ is
+ * removed.
+ *
+ * @module schemas/auth-tokens/sqlite
+ * @since v0.0.3-alpha (Plan A — schemas consolidation)
+ */
+
+import {
+  sqliteTable,
+  integer,
+  text,
+  index,
+  uniqueIndex,
+} from "drizzle-orm/sqlite-core";
+
+import { users } from "../users/sqlite";
+
+export const verificationTokens = sqliteTable(
+  "verification_tokens",
+  {
+    identifier: text("identifier").notNull(),
+    token: text("token").notNull(),
+    expires: integer("expires", { mode: "timestamp" }).notNull(),
+  },
+  t => [
+    uniqueIndex("verification_tokens_identifier_token_pk").on(
+      t.identifier,
+      t.token
+    ),
+    index("verification_tokens_token_idx").on(t.token),
+  ]
+);
+
+// Password reset tokens (custom table)
+export const passwordResetTokens = sqliteTable(
+  "password_reset_tokens",
+  {
+    id: integer("id").primaryKey({ autoIncrement: true }),
+    identifier: text("identifier").notNull(),
+    tokenHash: text("token_hash").notNull(),
+    expires: integer("expires", { mode: "timestamp" }).notNull(),
+    usedAt: integer("used_at", { mode: "timestamp" }),
+    createdAt: integer("created_at", { mode: "timestamp" })
+      .notNull()
+      .$defaultFn(() => new Date()),
+  },
+  t => [
+    uniqueIndex("prt_identifier_token_hash_unique").on(
+      t.identifier,
+      t.tokenHash
+    ),
+    index("prt_expires_idx").on(t.expires),
+    index("prt_used_at_idx").on(t.usedAt),
+  ]
+);
+
+// Email verification tokens (custom, hashed)
+export const emailVerificationTokens = sqliteTable(
+  "email_verification_tokens",
+  {
+    id: integer("id").primaryKey({ autoIncrement: true }),
+    identifier: text("identifier").notNull(),
+    tokenHash: text("token_hash").notNull(),
+    expires: integer("expires", { mode: "timestamp" }).notNull(),
+    createdAt: integer("created_at", { mode: "timestamp" })
+      .notNull()
+      .$defaultFn(() => new Date()),
+  },
+  t => [
+    uniqueIndex("evt_identifier_token_hash_unique").on(
+      t.identifier,
+      t.tokenHash
+    ),
+    index("evt_expires_idx").on(t.expires),
+  ]
+);
+
+// Refresh tokens for custom auth session management
+// Stores SHA-256 hashed opaque tokens, enables session revocation and token rotation
+export const refreshTokens = sqliteTable(
+  "refresh_tokens",
+  {
+    id: text("id").primaryKey(),
+    userId: text("user_id")
+      .notNull()
+      .references(() => users.id, { onDelete: "cascade" }),
+    // SHA-256 hex digest of the opaque refresh token (never store raw tokens)
+    tokenHash: text("token_hash").notNull(),
+    // Request metadata for session listing and security auditing
+    userAgent: text("user_agent"),
+    ipAddress: text("ip_address"),
+    expiresAt: integer("expires_at", { mode: "timestamp" }).notNull(),
+    createdAt: integer("created_at", { mode: "timestamp" })
+      .notNull()
+      .$defaultFn(() => new Date()),
+  },
+  t => [
+    index("refresh_tokens_token_hash_idx").on(t.tokenHash),
+    index("refresh_tokens_user_id_idx").on(t.userId),
+    index("refresh_tokens_expires_at_idx").on(t.expiresAt),
+  ]
+);
