+++ packages/nextly/src/schemas/index.ts
@@ -63,6 +63,14 @@ export * from "./_zod"; // Zod-only validators (user, rbac, validation)
 // Plan A Task 5 — user identity tables. PG re-exports here for direct-query
 // callers. Other dialects accessible via getCoreSchema(dialect).
 export { users, accounts, sessions } from "./users/postgres";
+
+// Plan A Task 6 — auth-token tables. PG re-exports for direct-query callers.
+export {
+  verificationTokens,
+  emailVerificationTokens,
+  passwordResetTokens,
+  refreshTokens,
+} from "./auth-tokens/postgres";
 export * from "./dynamic-collections"; // dialect-aware barrel — kept; unchanged
 export * from "./dynamic-components"; // kept; unchanged
 export * from "./migrations"; // dropped in Plan B; kept here in Plan A
